requirements.txt
getmac==0.9.4
paho-mqtt==2.0.0
